pyproject.toml
[tool.poetry]
name = "suzieq"
version = "0.20.0"
description = "A framework and application for network observability"
readme = 'README.md'
repository = 'https://github.com/netenglabs/suzieq'
documentation = 'https://suzieq.readthedocs.io'
authors = ['suzieq dev team']
maintainers = ['suzieq dev team']
keywords = ['network', 'Cisco', 'Arista', 'Junos', 'observability', 'Cumulus', 'monitoring', 'SoNIC', 'IOS']
license = 'Apache-2.0'
homepage = 'https://www.stardustsystems.net/suzieq/'
classifiers = [
   'Topic :: System :: Networking :: Monitoring'
]


[tool.poetry.dependencies]
python = ">3.8.1, < 3.10"
aiohttp = "==3.7.4"
asyncssh = "~2.11"
textfsm = "*"
async-timeout = "<4.0"
aiofiles = "*"
prompt-toolkit = ">2"
pyarrow = "10.0.1"
pandas = "1.3.4"
PyYAML = "*"
python-nubia = "0.2b5"
uvloop = "*"
networkx = "^2.4"
matplotlib = "^3.2.2"
faker = "^4.1.1"
netconan = "^0.11.2"
jsonpath-ng = "^1.5.1"
fastapi = "^0.65"
uvicorn = "^0.14.0"
tabulate = "^0.8.7"
streamlit = "1.11.1"
graphviz = "^0.15"
dateparser = "^1.1"
natsort = "^8.0.0"
colorama = "^0.4.4"
tzlocal = "< 3.0"
pyparsing = ">=2.2.0, <3.0"
xmltodict = "^0.12.0"
streamlit-aggrid = "^0.2.2"
ciscoconfparse = "^1.6.21"
notebook = "6.4.12"
urllib3 = "^1.26.12"
packaging = "^21.3"

[tool.poetry.dev-dependencies]
pylint = "*"
autopep8 = "*"
rope = "*"
jedi = "~0.17.2"
flake8 = "*"
ipdb = "*"
pycscope = "*"
pytest = "*"
pytest-cov = "*"
pytest-xdist = "*"
pytest-asyncio = "*"
pytest-mock = "*"
mock = "*"
pytest-depends = "^1.0.1"
bandit = "^1.6.2"
pytest-icdiff = "^0.5"
ipython = "~=7.16.3"
requests = "^2.24.0"
sphinx-autodoc-typehints = "^1.11.1"
filelock = "^3.0.12"
isort = "^5.10.1"
pre-commit = "^2.19.0"
toml = "^0.10.2"

[tool.poetry.scripts]
sq-poller = 'suzieq.poller.sq_poller:controller_main'
sq-rest-server = 'suzieq.restServer.sq_rest_server:rest_main'
sq-coalescer = 'suzieq.utilities.sq_coalescer:coalescer_main'
sq-anonymizer = 'suzieq.utilities.sq_anonymizer:anonymizer_main'
sq-simnodes = 'suzieq.utilities.sq_simnode:simnodes_main'
suzieq-cli = 'suzieq.cli.sq_cli:cli_main'
suzieq-gui = 'suzieq.gui.sq_gui:gui_main'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
